move query-specific methods into query object
